Refactored monolithic backend into multiple files
